feat: move all tiptap references from react lib to the editor lib
