+++ packages/editor/src/components/Editor/MathsModal.tsx
@@ -1,12 +1,11 @@
 import { useEffect, useState } from "react";
 
+import { Button, Modal } from "@edifice-ui/react";
 import { Editor, EditorContent, useEditor } from "@tiptap/react";
 import StarterKit from "@tiptap/starter-kit";
 import Mathematics from "@tiptap-pro/extension-mathematics";
 import { createPortal } from "react-dom";
 import { useTranslation } from "react-i18next";
-
-import { Button, Modal } from "../../components";
 
 interface ModalProps {
   isOpen: boolean;
