+++ packages/editor/src/components/NodeView/AttachmentNodeView.tsx
