+++ packages/editor/src/components/NodeView/ImageNodeView.tsx
