+++ packages/editor/src/components/NodeView/index.ts
@@ -3,4 +3,3 @@ export { default as AttachmentNodeView } from "./AttachmentNodeView";
 export { default as ImageNodeView } from "./ImageNodeView";
 export { default as LinkerNodeView } from "./LinkerNodeView";
 export { default as VideoNodeView } from "./VideoNodeView";
-export { useImageSelection } from "./useImageSelection";
