+++ packages/editor/src/components/Toolbar/TableToolbar.DelMenu.tsx
@@ -5,10 +5,9 @@ import {
   DeleteRow,
   DeleteRowHighlight,
 } from "@edifice-ui/icons";
+import { Dropdown } from "@edifice-ui/react";
 import { Editor } from "@tiptap/react";
 import { useTranslation } from "react-i18next";
-
-import { Dropdown } from "../../components";
 
 interface Props {
   /**
