+++ packages/editor/src/components/index.ts
@@ -1,2 +1,6 @@
 export * from "./Editor";
 export * from "./EditorToolbar";
+export * from "./BubbleMenuEditImage";
+export * from "./NodeView";
+export * from "./Renderer";
+export * from "./Toolbar";
