+++ packages/editor/src/hooks/index.ts
@@ -1,9 +1,11 @@
 export * from "./useActionOptions";
 export * from "./useEditorContext";
 export * from "./useImageModal";
+export * from "./useImageSelection";
 export * from "./useLinkToolbar";
 export * from "./useMathsModal";
 export * from "./useMediaLibraryModal";
+export * from "./useResizeMedia";
 export * from "./useSpeechRecognition";
 export * from "./useSpeechSynthetisis";
 export * from "./useTipTapEditor";
