+++ packages/editor/src/hooks/useImageSelection.ts
