+++ packages/react/package.json
@@ -46,22 +46,6 @@
     "@floating-ui/react": "0.26.0",
     "@pixi/react": "7.1.1",
     "@popperjs/core": "2.11.8",
-    "@tiptap-pro/extension-mathematics": "2.2.1",
-    "@tiptap/extension-color": "2.0.3",
-    "@tiptap/extension-highlight": "2.0.3",
-    "@tiptap/extension-table": "2.0.3",
-    "@tiptap/extension-table-cell": "2.0.3",
-    "@tiptap/extension-table-header": "2.0.3",
-    "@tiptap/extension-table-row": "2.0.3",
-    "@tiptap/extension-text-style": "2.0.3",
-    "@tiptap/extension-underline": "2.0.3",
-    "@tiptap/extension-image": "2.0.3",
-    "@edifice-tiptap-extensions/extension-attachment": "develop-b2school",
-    "@edifice-tiptap-extensions/extension-audio": "develop-b2school",
-    "@edifice-tiptap-extensions/extension-image": "develop-b2school",
-    "@edifice-tiptap-extensions/extension-linker": "develop-b2school",
-    "@edifice-tiptap-extensions/extension-table-cell": "develop-b2school",
-    "@edifice-tiptap-extensions/extension-video": "develop-b2school",
     "@tiptap/pm": "2.0.3",
     "@tiptap/react": "2.0.3",
     "@tiptap/starter-kit": "2.0.3",
