--- packages/react/src/tiptap/Modal/index.ts
@@ -1,1 +1,0 @@
-export { default as MathsModal } from "./MathsModal";
