--- packages/react/src/tiptap/RichTextEditor/TiptapWrapper.tsx
@@ -1,7 +1,0 @@
-import { ReactNode } from "react";
-
-const TiptapWrapper = ({ children }: { children: ReactNode }) => {
-  return <div className="border rounded">{children}</div>;
-};
-
-export default TiptapWrapper;
