--- packages/react/src/tiptap/RichTextEditor/index.ts
@@ -1,1 +1,0 @@
-export { default as TiptapWrapper } from "./TiptapWrapper";
