--- packages/react/src/tiptap/index.ts
@@ -1,6 +1,0 @@
-export * from "./Modal";
-export * from "./MediaWrapper";
-export * from "./NodeView";
-export * from "./Renderer";
-export * from "./RichTextEditor";
-export * from "./Toolbar";
